🚚 api: move controller method to src/certification/configuration context
