+++ api/src/certification/configuration/application/complementary-certification-controller.js
@@ -1,4 +1,5 @@
 import { usecases } from '../domain/usecases/index.js';
+import * as attachableTargetProfilesSerializer from '../infrastructure/serializers/attachable-target-profiles-serializer.js';
 import * as complementaryCertificationSerializer from '../infrastructure/serializers/complementary-certification-serializer.js';
 
 const findComplementaryCertifications = async function () {
@@ -6,7 +7,14 @@ const findComplementaryCertifications = async function () {
   return complementaryCertificationSerializer.serialize(complementaryCertifications);
 };
 
+const searchAttachableTargetProfilesForComplementaryCertifications = async function (request) {
+  const searchTerm = request.query.searchTerm;
+  const attachableTargetProfiles = await usecases.searchAttachableTargetProfiles({ searchTerm });
+  return attachableTargetProfilesSerializer.serialize(attachableTargetProfiles);
+};
+
 const complementaryCertificationController = {
   findComplementaryCertifications,
+  searchAttachableTargetProfilesForComplementaryCertifications,
 };
 export { complementaryCertificationController };
